requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==3.2.4
pip==22.0.2
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pytz==2023.3
setuptools==59.6.0
sqlparse==0.4.4
tomli==2.0.1
wheel==0.37.1
retrying==1.3.4
djangorestframework==3.14.0
django-cors-headers==4.0.0
six==1.16.0